Extract WarningsPanel class
Keeping the offsetting of line numbers out of WarningsPanel
to keep it simpler.
