+++ lib/Converter.js
@@ -1,6 +1,6 @@
 "use babel";
 import lebab from 'lebab';
-import {Point} from 'atom';
+import WarningsPanel from './WarningsPanel';
 
 export default class Converter {
     constructor() {
@@ -35,71 +35,9 @@ export default class Converter {
     }
 
     showWarnings(warnings) {
-        // Close warnings panel, if open from previous time
-        if (Converter.warningsPanel) {
-            this.hideWarnings();
-        }
-
-        if (warnings.length === 0) {
-            return;
-        }
-
-        // Store the panel in static property,
-        // so it will live through different Converter instances
-        Converter.warningsPanel = atom.workspace.addBottomPanel({
-            item: this.createWarningsContainer(warnings)
-        });
-    }
-
-    hideWarnings() {
-        Converter.warningsPanel.destroy();
-        Converter.warningsPanel = undefined;
-    }
-
-    createWarningsContainer(warnings) {
-        return this.createElement("div", "lebab-warnings-container",
-            [this.createCloseButton()].concat(
-                warnings.map(this.offsetLine, this).map(this.createWarning, this)
-            )
+        new WarningsPanel(this.editor).show(
+            warnings.map(this.offsetLine, this)
         );
-    }
-
-    createCloseButton() {
-        const link = this.createElement("a", "icon icon-x lebab-warnings-close");
-        link.onclick = () => {
-            this.hideWarnings();
-        };
-        return link;
-    }
-
-    createWarning({line, msg, type}) {
-        return this.createElement("p", "lebab-warning", [
-            this.createElement("span", "inline-block highlight", "Lebab"),
-            this.createElement("span", "inline-block highlight-warning", "Warning"),
-            this.createElement("span", "inline-block highlight", type),
-            this.createElement("span", "lebab-warning__msg", msg),
-            this.createLineNrLink(line),
-        ]);
-    }
-
-    createLineNrLink(line) {
-        const link = this.createElement("a", "lebab-warning__line", `at line ${line}`)
-        link.onclick = () => {
-            this.editor.setCursorBufferPosition(new Point(line - 1, 0));
-        };
-        return link;
-    }
-
-    createElement(nodeType, className, children) {
-        const el = document.createElement(nodeType);
-        el.className = className;
-        if (typeof children === "string") {
-            el.appendChild(document.createTextNode(children))
-        }
-        else if (Array.isArray(children)) {
-            children.forEach(child => el.appendChild(child));
-        }
-        return el;
     }
 
     offsetLine({line, msg, type}) {
